# internal/server/openapi/openapi-admin.yaml
openapi: 3.1.0
info:
  title: Firefly Admin API
  # Placeholder only. The running binary replaces this with its build version
  # at startup (server.SetBuildVersion); do not hand-edit it to track a release.
  version: "0.0.0-dev"
  description: >
    Administrative surface for the Firefly LLM gateway dashboard. Every
    endpoint requires an authenticated operator: either an active dashboard
    session token (obtained via `POST /api/auth/login`) or the configured
    admin bearer token, presented as `Authorization: Bearer <token>`.

    This document covers tenant lifecycle management under `/api/tenants`
    and provider/credential-pool administration under `/api/providers` and
    `/api/keys`. Tenant mutations persist through the same validate-then-store
    pipeline as the bulk settings endpoint (Turso when configured, otherwise
    the config files) and hot-swap the routing snapshot, so changes take effect
    immediately without a restart. Provider and key records live only in the
    Turso database and are refused with 503 when no store is configured. The
    public `/v1/*` surface is documented separately at `/api/openapi.yaml`.

    Credential discipline: stored keys are never returned. Key reads carry a
    masked `api_key_hint` only, and because that hint is the sole credential
    shape this surface ever shows, a secret in mask shape is refused on writes
    instead of being stored as a routable credential. `account_metadata` — the
...
  contact:
    name: Firefly
...
servers:
...
security:
  - AdminAuth: []
paths:
...
  /api/pricing/{key}:
    put:
      tags: [Pricing]
      parameters:
        - $ref: "#/components/parameters/PricingKey"
      summary: Create or replace one price entry
      description: >
        Upserts the entry for `key` (a model name, optionally a wildcard
        prefix). A new entry is appended; an existing one takes the new numbers
        and its `source` becomes `manual`, so a later models.dev import cannot
        silently revert an operator's edit. The sheet is persisted through the
        same pipeline as `POST /api/settings` and hot-swapped into the routing
        snapshot.
      operationId: upsertPricingEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingEntryUpsert"
      responses:
        "200":
          description: The stored entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  entry:
                    $ref: "#/components/schemas/PricingEntryUpsert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [Pricing]
      parameters:
        - $ref: "#/components/parameters/PricingKey"
      summary: Delete one price entry
      description: >
        Removes the entry for `key`. A model whose only price came from a
        models.dev seed simply becomes unpriced and falls back to the legacy
        flat rate, exactly as it did before any import.
      operationId: deletePricingEntry
      responses:
        "200":
          description: The entry was removed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  key:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: No entry exists for that key.
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      description: >
        Dashboard session token (from `POST /api/auth/login`) or the
        configured admin bearer token. A missing or invalid token returns 401.
  parameters:
...
    PricingKey:
      name: key
      in: path
      required: true
      description: >
        Price-sheet key: an exact model name or a wildcard prefix ending in
        `*` (the longest matching prefix wins). Matched against both the
        public model name and the upstream's private model name.
      schema:
        type: string
        examples: ["gpt-4o", "claude-sonnet-4-*", "openai/gpt-4o"]
  responses:
    Unauthorized:
      description: Missing or invalid admin credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              message: "unauthorized: valid dashboard session or admin token required"
              type: authentication_error
              param: null
              code: null
    BadRequest:
      description: Malformed JSON, a missing required field, a body/path name mismatch, or catalog validation failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: No configuration source (database, config files, or loaded snapshot) is available.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      description: OpenAI-shaped error envelope.
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message:
              type: string
            type:
              type: string
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found_error
                - rate_limit_error
                - api_error
            param:
              type: [string, "null"]
            code:
              type: [string, "null"]
...
    PricingEntryUpsert:
      type: object
      description: >
        One price-sheet entry. All prices are integer micro-USD per 1M tokens,
        so a provider's decimal per-1M price is stored exactly with no float
        rounding. `model` in the body is ignored — the path key wins.
      required: [input_micros_per_m, output_micros_per_m]
      properties:
        input_micros_per_m:
          type: integer
          format: int64
          minimum: 0
          description: Uncached prompt price per 1M tokens.
          examples: [2500000]
        output_micros_per_m:
          type: integer
          format: int64
          minimum: 0
          description: Completion price per 1M tokens.
          examples: [10000000]
        cache_read_micros_per_m:
          type: integer
          format: int64
          minimum: 0
          description: >
            Cached prompt read price per 1M tokens. When absent, cached tokens
            are billed at the input price rather than for free.
          examples: [1250000]
        cache_write_micros_per_m:
          type: integer
          format: int64
          minimum: 0
          description: Prompt cache write price per 1M tokens. Falls back to the input price when absent.
          examples: [3125000]
        source:
          type: string
          description: Ignored on write; an upsert always records `manual`.
          enum: [models.dev, manual]
        canonical_model_id:
          type: string
          description: >
            Provider-qualified id used to resolve a price for protocols that
            serve several providers (antigravity serves both Gemini and
            Claude).
          examples: ["openai/gpt-4o"]
...